#96: move tests to test bundle
