+++ tests/org.eclipse.n4js.hlc.tests/src/main/java/org/eclipse/n4js/hlc/tests/EnvironmentVariableUtils.java
@@ -8,7 +8,7 @@
  * Contributors:
  *   NumberFour AG - Initial API and implementation
  */
-package org.eclipse.n4js.hlc;
+package org.eclipse.n4js.hlc.tests;
 
 import org.apache.log4j.Logger;
