+++ tools/org.eclipse.n4js.hlc/pom.xml
@@ -155,30 +155,6 @@ Contributors:
 					<!-- successCodes>0</successCodes -->
 				</configuration>
 			</plugin>
-			<!-- possibly have to publish the artifact? -->
-			<!--  plugin>
-				<groupId>org.codehaus.mojo</groupId>
-				<artifactId>build-helper-maven-plugin</artifactId>
-				<version>1.9.1</version>
-				<executions>
-					<execution>
-						<id>attach-artifacts</id>
-						<phase>package</phase>
-						<goals>
-							<goal>attach-artifact</goal>
-						</goals>
-						<configuration>
-							<artifacts>
-								<artifact>
-									<file></file>
-									<type>extension of your file </type>
-									<classifier>optional</classifier>
-								</artifact>
-							</artifacts>
-						</configuration>
-					</execution>
-				</executions>
-			</plugin-->
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-compiler-plugin</artifactId>
@@ -188,47 +164,7 @@ Contributors:
 					<target>${jdk-version}</target>
 				</configuration>
 			</plugin>
-			<!-- Publish test-jar with shared test-helper classes -->
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-jar-plugin</artifactId>
-<!-- 					<configuration> -->
-<!-- 						<systemProperties> -->
-<!-- 							<NODEJS_PATH>${NODEJS_PATH}</NODEJS_PATH> -->
-<!-- 							<org.eclipse.n4js.defaultNodePath>${org.eclipse.n4js.defaultNodePath}</org.eclipse.n4js.defaultNodePath> -->
-<!-- 						</systemProperties> -->
-<!-- 					</configuration> -->
-				<version>2.6</version>
-				<executions>
-					<execution>
-						<goals>
-							<goal>test-jar</goal>
-						</goals>
-						<configuration>
-							<includes>
-								<include>**/*</include>
-							</includes>
-						</configuration>
-					</execution>
-				</executions>
-			</plugin>
 
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-surefire-plugin</artifactId>
-				<version>${maven-surefire-plugin.version}</version>
-				<executions>
-					<execution>
-						<id>default-test</id>
-						<configuration>
-							<systemProperties>
-								<NODEJS_PATH>${NODEJS_PATH}</NODEJS_PATH>
-								<org.eclipse.n4js.defaultNodePath>${org.eclipse.n4js.defaultNodePath}</org.eclipse.n4js.defaultNodePath>
-							</systemProperties>
-						</configuration>
-					</execution>
-				</executions>
-			</plugin>
 
 			<!--
 			The following will deploy this artifact to a temporary maven repository (aka "local build repo").
@@ -289,86 +225,7 @@ Contributors:
 		</pluginManagement>
 	</build>
 
-	<profiles>
-		<profile>
-			<id>noTests</id>
-			<activation>
-				<property>
-					<name>noTests</name>
-				</property>
-			</activation>
-			<build>
-				<plugins>
-					<plugin>
-						<groupId>org.apache.maven.plugins</groupId>
-						<artifactId>maven-compiler-plugin</artifactId>
-						<executions>
-							<execution>
-								<id>default-testCompile</id>
-								<phase>none</phase>
-								<configuration>
-									<systemProperties>
-										<NODEJS_PATH>${NODEJS_PATH}</NODEJS_PATH>
-										<org.eclipse.n4js.defaultNodePath>${org.eclipse.n4js.defaultNodePath}</org.eclipse.n4js.defaultNodePath>
-									</systemProperties>
-								</configuration>
-							</execution>
-						</executions>
-					</plugin>
-					<plugin>
-						<groupId>org.apache.maven.plugins</groupId>
-						<artifactId>maven-resources-plugin</artifactId>
-						<version>${maven-resources-plugin.version}</version>
-						<executions>
-							<execution>
-								<id>default-testResources</id>
-								<phase>none</phase>
-								<configuration>
-									<systemProperties>
-										<NODEJS_PATH>${NODEJS_PATH}</NODEJS_PATH>
-										<org.eclipse.n4js.defaultNodePath>${org.eclipse.n4js.defaultNodePath}</org.eclipse.n4js.defaultNodePath>
-									</systemProperties>
-								</configuration>
-							</execution>
-						</executions>
-					</plugin>
-					<plugin>
-						<groupId>org.apache.maven.plugins</groupId>
-						<artifactId>maven-surefire-plugin</artifactId>
-						<version>${maven-surefire-plugin.version}</version>
-						<executions>
-							<execution>
-								<id>default-test</id>
-								<phase>none</phase>
-								<configuration>
-									<systemProperties>
-										<NODEJS_PATH>${NODEJS_PATH}</NODEJS_PATH>
-										<org.eclipse.n4js.defaultNodePath>${org.eclipse.n4js.defaultNodePath}</org.eclipse.n4js.defaultNodePath>
-									</systemProperties>
-								</configuration>
-							</execution>
-						</executions>
-					</plugin>
-					<plugin>
-						<groupId>org.codehaus.mojo</groupId>
-						<artifactId>exec-maven-plugin</artifactId>
-						<executions>
-							<execution>
-								<id>it-of-n4js-compiler-command</id>
-								<phase>none</phase>
-								<configuration>
-									<systemProperties>
-										<NODEJS_PATH>${NODEJS_PATH}</NODEJS_PATH>
-										<org.eclipse.n4js.defaultNodePath>${org.eclipse.n4js.defaultNodePath}</org.eclipse.n4js.defaultNodePath>
-									</systemProperties>
-								</configuration>
-							</execution>
-						</executions>
-					</plugin>
-				</plugins>
-			</build>
-		</profile>
-	</profiles>
+
 
 
 
@@ -600,17 +457,6 @@ Contributors:
 			<groupId>com.google.guava</groupId>
 			<artifactId>guava</artifactId>
 			<version>[15.0.0]</version>
-		</dependency>
-
-		<!-- # # Below are dependencies which don't go into the jar # # # # # # 
-			# # # # # # # # # # # # # # # # # # They are required for testing. -->
-		<!-- JUnit is required for compiling the tests It will not be copied to 
-			the jar -->
-		<dependency>
-			<groupId>junit</groupId>
-			<artifactId>junit</artifactId>
-			<version>4.12</version>
-			<scope>test</scope>
 		</dependency>
 	</dependencies>
 </project>
